Install 3rd party tools
This is a bit of an all-in-one type installer for external programs that
should be installed on a machine. It covers Apt, Homebrew, Git repos,
and general bash-script installers (e.g. run this curl command...).

This should eventually be incorporated into the entry-level installer
menu, but this is the first pass.