Install Chocolatey packages
Make use of modules rather than plain scripts as it allows exporting the same two function names for every utility.
This way there is a way to test if it is already installed independent of the package name.
This is especially useful for those packages that have both a '.install' and '.portable' variation.